requirements.txt
supabase
python-dotenv
fastapi
uvicorn[standard]